+++ src/main/java/de/heinrich/spoc/domain/Material.java
@@ -19,22 +19,22 @@ public class Material implements Serializable {
     )
     private Long id;
     private String a_name;
-    private BigDecimal fossiles;
-    private BigDecimal prozessenergie;
-    private BigDecimal productionCO2;
-    private BigDecimal bioco2prod;
-    private BigDecimal bioFuelCO2;
-    private BigDecimal dichte;
-    private BigDecimal co2Verbrennung;
-    private BigDecimal bioCO2Verbrennung;
-    private BigDecimal heizenergie;
-    private BigDecimal co2Recycling;
+    private double fossiles;
+    private double prozessenergie;
+    private double productionCO2;
+    private double bioco2prod;
+    private double bioFuelCO2;
+    private double dichte;
+    private double co2Verbrennung;
+    private double bioCO2Verbrennung;
+    private double heizenergie;
+    private double co2Recycling;
 
-    public BigDecimal getCo2Recycling() {
+    public double getCo2Recycling() {
         return co2Recycling;
     }
 
-    public void setCo2Recycling(BigDecimal co2recycling) {
+    public void setCo2Recycling(double co2recycling) {
         this.co2Recycling = co2recycling;
     }
 
@@ -46,67 +46,67 @@ public class Material implements Serializable {
         this.a_name = name;
     }
 
-    public BigDecimal getFossiles() {
+    public double getFossiles() {
         return fossiles;
     }
 
-    public void setFossiles(BigDecimal fossiles) {
+    public void setFossiles(double fossiles) {
         this.fossiles = fossiles;
     }
 
-    public BigDecimal getProzessenergie() {
+    public double getProzessenergie() {
         return prozessenergie;
     }
 
-    public void setProzessenergie(BigDecimal prozessenergie) {
+    public void setProzessenergie(double prozessenergie) {
         this.prozessenergie = prozessenergie;
     }
 
-    public BigDecimal getProductionCO2() {
+    public double getProductionCO2() {
         return productionCO2;
     }
 
-    public void setProductionCO2(BigDecimal productionCO2) {
+    public void setProductionCO2(double productionCO2) {
         this.productionCO2 = productionCO2;
     }
 
-    public BigDecimal getBioFuelCO2() {
+    public double getBioFuelCO2() {
         return bioFuelCO2;
     }
 
-    public void setBioFuelCO2(BigDecimal bioFuel) {
+    public void setBioFuelCO2(double bioFuel) {
         this.bioFuelCO2 = bioFuel;
     }
 
-    public BigDecimal getDichte() {
+    public double getDichte() {
         return dichte;
     }
 
-    public void setDichte(BigDecimal dichte) {
+    public void setDichte(double dichte) {
         this.dichte = dichte;
     }
 
-    public BigDecimal getCo2Verbrennung() {
+    public double getCo2Verbrennung() {
         return co2Verbrennung;
     }
 
-    public void setCo2Verbrennung(BigDecimal co2Verbrennung) {
+    public void setCo2Verbrennung(double co2Verbrennung) {
         this.co2Verbrennung = co2Verbrennung;
     }
 
-    public BigDecimal getBioCO2Verbrennung() {
+    public double getBioCO2Verbrennung() {
         return bioCO2Verbrennung;
     }
 
-    public void setBioCO2Verbrennung(BigDecimal bioCO2Verbrennung) {
+    public void setBioCO2Verbrennung(double bioCO2Verbrennung) {
         this.bioCO2Verbrennung = bioCO2Verbrennung;
     }
 
-    public BigDecimal getHeizenergie() {
+    public double getHeizenergie() {
         return heizenergie;
     }
 
-    public void setHeizenergie(BigDecimal heizenergie) {
+    public void setHeizenergie(double heizenergie) {
         this.heizenergie = heizenergie;
     }
 
@@ -118,11 +118,11 @@ public class Material implements Serializable {
         this.a_name = a_name;
     }
 
-    public BigDecimal getBioco2prod() {
+    public double getBioco2prod() {
         return bioco2prod;
     }
 
-    public void setBioco2prod(BigDecimal bioco2prod) {
+    public void setBioco2prod(double bioco2prod) {
         this.bioco2prod = bioco2prod;
     }
